[REFACTOR] CoinDetail API 통신 리팩토링 및 차트 화면 리팩토링

+++ SeSAC_CoingeckoProject/NumberFormatManager.swift
@@ -19,7 +19,7 @@ class NumberFormatManager {
     // 이렇게 작성하면 한 번 클래스를 만든 이상 내려가지 않을 것이다. 계속쓸 예정일 때 적합
     private let numberFormatter = NumberFormatter()
     
-    func calculator(_ number: Int) -> String {
+    func calculator(_ number: Double) -> String {
         numberFormatter.numberStyle = .decimal// 3자리마다 쉼표 찍어주는 기능
         let result = numberFormatter.string(from: number as NSNumber)
         // result는 String?타입
